+++ CHANGES.rst
@@ -12,6 +12,7 @@ Changelog
 
 **Fixed**
 
+- #452 Traceback error when deleting attachment from Analysis Request
 - #450 Traceback after clicking "Manage Results" in a WS w/o Analyses assigned
 - #445 Fix AR Add Form: No sample points are found if a sample type was set